Revert "Add ClientProxy"
This reverts commit 953b65ffe6bbc81ac451accd75c101720743cdb9.
