+++ src/main/java/tconstruct/tools/items/TFManualInfo.java
@@ -9,7 +9,7 @@ import cpw.mods.fml.common.FMLCommonHandler;
 import cpw.mods.fml.relauncher.Side;
 import mantle.books.BookData;
 import mantle.books.BookDataStore;
-import tconstruct.client.TFTProxyClient;
+import twilightforest.integration.TFTinkerConstructIntegration;
 
 public class TFManualInfo {
 
@@ -21,7 +21,7 @@ public class TFManualInfo {
                 twilightMaterials,
                 "tconstruct.manual.twilightmaterials",
                 "\u00a7o" + StatCollector.translateToLocal("manualTF.tooltip"),
-                side == Side.CLIENT ? TFTProxyClient.twilightMaterials : null,
+                side == Side.CLIENT ? TFTinkerConstructIntegration.twilightMaterials : null,
                 "tinker:tinkerbook_diary");
     }
